Refactored code and fixes styling issues
