+++ src/components/Message/index.jsx
@@ -1,7 +1,7 @@
 export default function Message({ winner, children }) {
   return (
     <>
-      <div className="text">{!winner ? "Tie" : "Won:"}</div>
+      <div className="text">{winner === false ? "Tie" : "Won:"}</div>
       <header>{children}</header>
     </>
   );
