+++ src/index.css
@@ -32,24 +32,24 @@ body {
 
 .board {
   width: fit-content;
-  margin: 40px auto;
+  margin: 100px auto;
   text-align: center;
+  justify-content: center;
 }
 
 .board h1 {
   color: #eee;
-  margin-bottom: 16px;
+  margin-bottom: 25px;
 }
 
 .board button {
   padding: 8px 12px;
-  margin: 25px;
   background: transparent;
   border: 2px solid #eee;
   color: #eee;
   width: 100px;
   border-radius: 5px;
-  transition: 0.2s;
+  transition: 0.5s ease-in-out;
   font-weight: bold;
   cursor: pointer;
 }
@@ -62,7 +62,7 @@ body {
 .board .game {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
-  gap: 10px;
+  gap: 5px;
 }
 
 .turn {
